# [SLA][TR] PIM-5208: patch for grid performance with large amount of attributes