--- a/CHANGELOG-1.4.md
+++ b/CHANGELOG-1.4.md
@@ -1,9 +1,10 @@
 # 1.4.x
 
 ## Bug fixes
-- PIM-5238: fix scroll on multiselect for mass edit
-- PIM-5177: fix login redirection
+- PIM-5238: Fix scroll on multiselect for mass edit
+- PIM-5177: Fix login redirection
 - PIM-5235: Fix empty reference data name on attributes import
+- PIM-5208: Fix the datagrid performance issue related to large number of attributes, only attribute usable in grid will be available
 
 # 1.4.11 (2015-11-27)