TEST PROTOCOL:
* no xdebug, done on local machine
* database with 10k attributes
* 1 variant group,  1 product that belongs to that variant group
* calculate mean time of  25 curl calls

BEFORE PATCH:
* refresh product grid: 9.278s
* refresh variant product grid: 9.235s

AFTER PATCH:
* refresh product grid: 0.439s
* refresh variant product grid: 0.453s

| Q                 | A
| ----------------- | ---
| Specs             | -
| Behats            | Y
| Blue CI           | CE Y, EE running
| Changelog updated | Y
| Review and 2 GTM  |

- [x] Update translation in attribute configuration to 'Useable in grid'
- [x] Add chapter in the http://docs.akeneo.com/latest/reference/performances_guide/index.html. PR here https://github.com/akeneo/pim-docs/pull/169/files
- [x] Add custom logic for variant group products grid to allow to display attribue axis values and filter by attribute axis (add a specific context configurator for group grid to always append axis attributes?)