@nidup 2015-11-21
really reduces the issue with attribute loading + metadata preparation but has other functional and technical impacts, so more code to come ;) (cf related ticket in jira)