@Nuscly 2015-11-23
We need a validator to check the amount of attributes as grid filter to avoid misconfiguration